requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
langchain_anthropic
langchain-google-genai

# New dependencies for crypto and MCP support
aiohttp
fastmcp
uvicorn
textblob
python-dotenv
ccxt
binance-python

# LM Studio local model integration
openai>=1.0.0  # For OpenAI-compatible API client
httpx  # Async HTTP client for better performance
